feat: refactor package structure for alcohol-related tests and controllers
